Moved UploadModal from document page to DragAndDropBox
